refactor: consolidate validation utilities into validator/common.py
- Create validator/common.py with shared BUILTIN_NAMES, PYTHON_KEYWORDS,
  and extract_names_from_expression utilities
- Move hydration validation functions to validator/hydration.py
- Remove duplicate validation code from hydrator_utils.py (~370 lines)
- Update syntactic.py to import from common.py
- Update hydrator.py imports to use validator module
